Totem abstraction: fix transition to gst_plugin_scanner profile.
Apparently the behavior of "Cix -> profile" has changed; I think it used to
(incorrectly?) accept to transition to a non-child profile whose name didn't
match the executable name, and now seems to be simply ignored. I'll consider
this as a bugfix. Let's use px instead, which works and matches more closely
what we want here.